+++ sdl_window.cpp
@@ -80,9 +80,9 @@ void SdlWindow::init_video( int& res_width
 
 	atexit(SDL_Quit);
 
-	const uint32_t requested_size = res_width * res_height;
-	uint32_t closest_size_diff = std::numeric_limits<uint32_t>::max();
-	uint16_t final_res_w = res_width, final_res_h = res_height;
+	const int requested_size = res_width * res_height;
+	int closest_size_diff = std::numeric_limits<int>::max();
+	int final_res_w = res_width, final_res_h = res_height;
 
     for(int display_mode_i = 0; display_mode_i < SDL_GetNumDisplayModes(0); ++display_mode_i)
     {
@@ -93,7 +93,7 @@ void SdlWindow::init_video( int& res_width
             return;
         }
 
-        const uint32_t size_diff = requested_size + mode.w*mode.h - 2*min(int(res_width), mode.w) * min(int(res_height), mode.h);
+        const int size_diff = requested_size + mode.w*mode.h - 2*min(res_width, mode.w) * min(res_height, mode.h);
 		if(size_diff < closest_size_diff)
 		{
 			closest_size_diff = size_diff;
